Move dashTest transaction test to forks/
