+++ test/transaction.js
@@ -2,9 +2,7 @@
 
 var assert = require('assert')
 var bscript = require('../src/script')
-var networks = require('../src/networks')
 var fixtures = require('./fixtures/transaction')
-var fixturesDashTest = require('./fixtures/forks/dashTest/transaction')
 var Transaction = require('../src/transaction')
 
 describe('Transaction', function () {
@@ -86,49 +84,6 @@ describe('Transaction', function () {
       var tx = Transaction.fromHex(txHex)
       assert.equal(-1, tx.version)
       assert.equal(0xffffffff, tx.locktime)
-    })
-  })
-
-  describe('fromBuffer/fromHex for Testnet Dash', function () {
-    fixturesDashTest.valid.forEach(function (testData) {
-      it('imports ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.dashTest)
-        assert.equal(tx.version, testData.version)
-        if (tx.versionSupportsDashSpecialTransactions()) {
-          assert.equal(tx.type, testData.type)
-        }
-        assert.equal(tx.locktime, testData.locktime)
-        assert.equal(tx.ins.length, testData.vin.length)
-        assert.equal(tx.outs.length, testData.vout.length)
-        if (tx.isDashSpecialTransaction()) {
-          assert.equal(tx.extraPayload.toString('hex'), testData.extraPayload)
-        }
-      })
-    })
-
-    fixturesDashTest.valid.forEach(function (testData) {
-      it('exports ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.dashTest)
-        const hexTx = tx.toHex()
-        assert.equal(testData.hex, hexTx)
-      })
-    })
-
-    fixturesDashTest.valid.forEach(function (testData) {
-      it('clone ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.dashTest)
-        const clonedTx = tx.clone()
-        assert.equal(clonedTx.toHex(), testData.hex)
-      })
-    })
-  })
-
-  describe('getPrevoutHash', function () {
-    fixturesDashTest.valid.filter(f => !!f.proRegTx).forEach(function (testData) {
-      it('produces the correct inputsHash on ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.dashTest)
-        assert.equal(tx.getPrevoutHash(Transaction.SIGHASH_ALL).reverse().toString('hex'), testData.proRegTx.inputsHash)
-      })
     })
   })
